// apps/frontend/src/infrastructure/api/types/list-markets-params.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Grocery Store API
 * API de gerenciamento de compras
 * OpenAPI spec version: 1.0.0
 */
import type { ListMarketsOrderBy } from './list-markets-order-by';
import type { ListMarketsOrderDirection } from './list-markets-order-direction';

export type ListMarketsParams = {
  search?: string;
  /**
   * @minimum 0
   */
  pageIndex?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  pageSize?: number;
  orderBy?: ListMarketsOrderBy;
  orderDirection?: ListMarketsOrderDirection;
  location?: {
    latitude?: number;
    longitude?: number;
  };
  expand?: boolean | string;
};